requirement.txt
cffi==1.12.3
setuptools==40.8.0
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycparser==2.19
Werkzeug==0.16.0
clipspy==0.3.2
